Working on fixing the typedef variable for the DWARF analysis; To-do: Uncomment necessary codes
The objective of this refactoring is to make the typedef variable handle the struct data. For example, in the case of typedef struct http_request (in the tiny.c), need to first create typedef with the name of http_request and then make it contain struct members for the http_request. Afterwards, when the DW_TAG_variable is handled, it will first check whether the typedef exists and then load the typedef struct information to recalculate the offset based on the variable's location