Updated the code in the analyze_var to support the latest pyelftools (0.31 now); To-do: Support inline function
Currently, dwarf analysis does not support the inlined function (e.g., how a function is first defined, and then called
in the main function). For the inlined function, there will be two DW_TAG_subprogram where one instance will contain the
location information, but the other instance will contain the other information such as name. In order to support this,
we may need to implement the subprogram class. Variables offset location with respect to the CFA may need to be calculated
first, and then when the subprogram is finalized, update those variables offset with the updated CFA base address.